Move ModelBase to common.database
