+++ starlette_web/common/database/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 
 from starlette_web.common.database.columns import EnumTypeColumn
+from starlette_web.common.database.model_base import ModelBase
 from starlette_web.common.database.model_mixin import ModelMixin, DBModel
 from starlette_web.common.database.session_maker import make_session_maker
 from starlette_web.common.database.types import ChoiceType
