+++ starlette_web/common/database/model_base.py
@@ -1,5 +1,4 @@
 from sqlalchemy.orm import declarative_base
 
 
-# TODO: maybe move to common.database ?
 ModelBase = declarative_base()
